requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
